Added win/loss updates
Winning or losing a battle (except by closing the webpage, which I can't do anything about) now updates the stored battle history, which is displayed in the popup. The issue is that finishing a battle in any format clears all the other formats' history for that opponent, due to how setting object key-values works. Will fix next commit.